Moved stuff around, Added a spinner on loading owner data
