+++ src/app/components/owner-view/financial-report/financial-report.component.scss
@@ -5,7 +5,7 @@ canvas {
   position: relative;
   max-width: 100%;
   max-height: 100%;
-  overflow: scroll;
+  overflow: auto;
 }
 .finances-card {
   position: relative;
