+++ src/app/components/owner-view/general-stats/general-stats.component.html
