+++ src/app/components/owner-view/general-stats/general-stats.component.scss
@@ -1,5 +1,5 @@
 #table-container {
   position: relative;
   max-height: 100%;
-  overflow-y: scroll;
+  overflow-y: auto;
 }
