+++ src/app/components/owner-view/owner-view.component.scss
@@ -1,0 +1,14 @@
+.layoutBox-v {
+  margin: 1%;
+}
+.layoutBox-percentages {
+  height: 150%;
+}
+
+.spinner {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
