+++ src/app/components/owner-view/percentage-indicator/percentage-indicator.component.html
@@ -6,7 +6,7 @@
     <div style="display: block; height: 80px; width: 100%; padding: 10px">
       <mat-progress-spinner
         class="percentage-margin"
-        color="primary"
+        color="accent"
         [mode]="mode"
         [value]="100"
         style="position: absolute; bottom: 5%"
