+++ src/app/components/owner-view/percentage-indicator/percentage-indicator.component.spec.ts
