--- src/app/owner-view/owner-view.component.scss
@@ -1,6 +1,0 @@
-.layoutBox-v {
-  margin: 1%;
-}
-.layoutBox-percentages {
-  height: 150%;
-}
